Add HTML5 export
This has been tested locally with a gamepad and seems to work. However,
because audio won't play until the screen is clicked on, a button was
also added to SplashScreen that only shows up in the Web build. Similarly,
the quit button is not shown on the web, because you cannot 'quit'
from an open browser tab.